+++ tests/VectorException/enc/enc.c
@@ -24,7 +24,7 @@ int DivideByZeroExceptionFunction(void)
     volatile float f = 0;
     volatile double d = 0;
 
-    f = 0.31;
+    f = 0.31f;
     d = 0.32;
 
     // Using inline assembly for idiv to prevent it being optimized out
@@ -125,7 +125,7 @@ static oe_vectored_exception_handler_t g_test_div_by_zero_handler;
 
 int VectorExceptionSetup()
 {
-    uint64_t ret = -1;
+    int ret = -1;
     oe_result_t result;
 
     // Add one exception handler.
